Allow custom environments or other setup scripts to bind to Circus events.
