+++ packages/jest-circus/src/state.ts
@@ -5,44 +5,55 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import {Event, State, EventHandler, STATE_SYM} from './types';
+import {
+  Event,
+  State,
+  EventHandler,
+  STATE_SYMBOL,
+  JEST_EVENT_HANDLERS_SYMBOL,
+} from './types';
 
 import {makeDescribe} from './utils';
 import eventHandler from './eventHandler';
 import formatNodeAssertErrors from './formatNodeAssertErrors';
 
-const eventHandlers: Array<EventHandler> = [
-  eventHandler,
-  formatNodeAssertErrors,
-];
-
 export const ROOT_DESCRIBE_BLOCK_NAME = 'ROOT_DESCRIBE_BLOCK';
 
-const ROOT_DESCRIBE_BLOCK = makeDescribe(ROOT_DESCRIBE_BLOCK_NAME);
-const INITIAL_STATE: State = {
-  currentDescribeBlock: ROOT_DESCRIBE_BLOCK,
-  currentlyRunningTest: null,
-  expand: undefined,
-  hasFocusedTests: false, // whether .only has been used on any test/describe
-  includeTestLocationInResult: false,
-  parentProcess: null,
-  rootDescribeBlock: ROOT_DESCRIBE_BLOCK,
-  testNamePattern: null,
-  testTimeout: 5000,
-  unhandledErrors: [],
+const getEventHandlers = () => {
+  if (!global[JEST_EVENT_HANDLERS_SYMBOL]) {
+    global[JEST_EVENT_HANDLERS_SYMBOL] = [eventHandler, formatNodeAssertErrors];
+  }
+  return global[JEST_EVENT_HANDLERS_SYMBOL];
 };
 
-global[STATE_SYM] = INITIAL_STATE;
+export const getState = (): State => {
+  if (!global[STATE_SYMBOL]) {
+    const ROOT_DESCRIBE_BLOCK = makeDescribe(ROOT_DESCRIBE_BLOCK_NAME);
+    global[STATE_SYMBOL] = {
+      currentDescribeBlock: ROOT_DESCRIBE_BLOCK,
+      currentlyRunningTest: null,
+      expand: undefined,
+      hasFocusedTests: false, // whether .only has been used on any test/describe
+      includeTestLocationInResult: false,
+      parentProcess: null,
+      rootDescribeBlock: ROOT_DESCRIBE_BLOCK,
+      testNamePattern: null,
+      testTimeout: 5000,
+      unhandledErrors: [],
+    } as State;
+  }
 
-export const getState = (): State => global[STATE_SYM];
-export const setState = (state: State): State => (global[STATE_SYM] = state);
+  return global[STATE_SYMBOL];
+};
+
+export const setState = (state: State): State => (global[STATE_SYMBOL] = state);
 
 export const dispatch = (event: Event): void => {
-  for (const handler of eventHandlers) {
+  for (const handler of getEventHandlers()) {
     handler(event, getState());
   }
 };
 
 export const addEventHandler = (handler: EventHandler): void => {
-  eventHandlers.push(handler);
+  getEventHandlers().push(handler);
 };
